Language Detection Problem
There are (N) predefined languages, each containing a set words. I determine which of these languages a given sentence belongs to, and output the index of the language (starting from 1) that have the most number of words in the given sentences, if there is more than one answer print all of them (in ascending order). All the words in the given languages are unique. Moreover, it is guarantee that the given sentence contain at least one word from one of the given languages.